requirements.txt
yacs
torch==1.9.0
cudatoolkit==10.2
numpy
matplotlib
torchvision==0.10.0
h5py
pickle-mixin
pillow
tqdm
pytorch3d
regex
requests
easydict
ninja
cython
dill
opencv-python
pycocotools
cffi
scipy
msgpack
pyaml
tensorboardX
timm